refactor(040): migra packages/config para TypeScript
