+++ packages/storage/src/index.js
@@ -12,7 +12,7 @@
  *   const json = await getJSON(adapter, 'jsonKey', defaultValue)
  */
 
-export { assertStorageAdapter } from './contracts.js'
+export { assertStorageAdapter } from './contracts'
 export { createWebStorageAdapter } from './webStorage.js'
 export { createMemoryStorageAdapter } from './memoryStorage.js'
 export { getJSON, setJSON } from './json.js'
